made depth sensor 20hz
